pyproject.toml
[tool.poetry]
name = "eqatlas"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.95.0"
mangum = "^0.17.0"
pycryptodome = "^3.17"
requests = "^2.28.2"
beautifulsoup4 = "^4.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
